Added Problem 3 Strawberry
Programming for Beginners – 24 April 2016
Problem 3. Strawberry
Your task is to write a program, which reads a number N from the console
and draws a strawberry, which size depends on N. See the examples below
to understand your task better.
Input
On the only input line you will be given the positive number N – odd
number in the range [1 … 17]
Output
The output should be printed on the console. Use the “#” (number sign)
for the outlines of the strawberry, “.” for the inside of the
strawberry. For the leaves:  left ones – “\”, middle ones – “|” and
right ones – “/”
Examples
Input Output Input Output Input Output
5 \-----|-----/
--\---|---/--
-----#.#-----
---#.....#---
-#.........#-
#...........#
-#.........#-
--#.......#--
---#.....#---
----#...#----
-----#.#----- 7 \-------|-------/
--\-----|-----/--
----\---|---/----
-------#.#-------
-----#.....#-----
---#.........#---
-#.............#-
#...............#
-#.............#-
--#...........#--
---#.........#---
----#.......#----
-----#.....#-----
------#...#------
-------#.#------- 9 \---------|---------/
--\-------|-------/--
----\-----|-----/----
------\---|---/------
---------#.#---------
-------#.....#-------
-----#.........#-----
---#.............#---
-#.................#-
#...................#
-#.................#-
--#...............#--
---#.............#---
----#...........#----
-----#.........#-----
------#.......#------
-------#.....#-------
--------#...#--------
---------#.#---------